Move API key help to a step-by-step guide and split components
