+++ cda-gui/src/main.jsx
@@ -24,6 +24,7 @@ import Timestamps from "./pages/timestamps";
 import LegacyFormat from "./pages/legacy-format/index.jsx";
 import UserLists from "./pages/user-lists/index.jsx";
 import ApiKeys from "./pages/api-keys/index.jsx";
+import ApiKeyHelp from "./pages/api-keys/help.jsx";
 import { routePaths } from "./route-paths";
 import AppAuthProvider from "./components/AppAuthProvider.jsx";
 import GlobalErrorBoundary from "./components/GlobalErrorBoundary.jsx";
@@ -40,6 +41,7 @@ const routeComponents = {
   "location-search": LocationSearch,
   "user-lists": UserLists,
   "api-keys": ApiKeys,
+  "api-key-help": ApiKeyHelp,
 };
 
 const router = createBrowserRouter(
