+++ cda-gui/src/pages/api-keys/README.md
@@ -4,6 +4,15 @@ Signed-in users can open `/cwms-data/api-keys` from the API Keys navigation link
 or from any of the four Authorization key operations in Swagger. Signed-out
 visitors are redirected to Home, and the navigation link is hidden.
 
+`/cwms-data/api-keys/help` is a separate signed-in guide with four numbered
+steps: create, save, send a request, and replace or revoke. It preserves office
+context when returning to key management and has a separate troubleshooting
+section. Both routes are excluded from the public sitemap.
+
+Page entrypoints compose the files in `components/`: authentication guard,
+manager, header, office context, key list/details, individual dialogs, and guide
+sections. API requests and date handling remain in `api.js`.
+
 The page uses Groundwork controls and dialogs, Groundwork Water's authentication
 and profile provider, and the existing `cwmsjs` Authorization API. No unpublished
 shared-library changes are required. Its card layout follows the User Roles page
@@ -32,8 +41,10 @@ The client still handles request serialization and URL encoding.
 - `gradlew :cwms-data-api:test --tests '*SpaErrorStatusFilterTest'` verifies the
   server's direct page route, including the trailing slash.
 - Browser checks with mocked CDA responses: signed-in list/detail, office
-  context, help, create/one-time secret, storage checks, revoke cancellation and
-  success, 403 recovery, signed-out direct route/navigation, and 390px layout.
+  context, guide navigation/direct reload and return-office selection,
+  create/one-time secret, storage checks, revoke cancellation and
+  success, 403 recovery, signed-out direct routes/navigation, and 390px layout
+  including wrapped code examples.
   Screenshots contain mock users and key names, with no secrets.
 
 Live database key operations and a deployed WAR still need integration testing.
