+++ cda-gui/src/pages/api-keys/help.jsx
@@ -1,0 +1,10 @@
+import RequireSignIn from "./components/RequireSignIn";
+import KeyHelpGuide from "./components/KeyHelpGuide";
+
+export default function ApiKeyHelp() {
+  return (
+    <RequireSignIn>
+      <KeyHelpGuide />
+    </RequireSignIn>
+  );
+}
